Fix perform:withArguments: primitive
