+++ core-lib/TestSuite/ReflectionTests.som
@@ -1,5 +1,4 @@
-"
-
+(*
 $Id: ReflectionTest.som 30 2009-07-31 12:20:25Z michael.haupt $
 
 Copyright (c) 2007-2013 see AUTHORS file
@@ -22,44 +21,54 @@ FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
-"
+THE SOFTWARE.*)
+class ReflectionTest usingPlatform: platform testFramework: minitest = (
+| private TestContext = minitest TestContext. 
+  private system = platform system.
+  private platform = platform.
+  private Array     = platform kernel Array.
+  private ObjectMirror = platform mirrors ObjectMirror.
+|
+) (
+  public class FactoryMethodTest = TestContext ()(
+    public class ReflectionTarget = (| num ::= 0. |)(
+      public set: val = (num := val)
+      public get = (^ num)
+    )
 
-ReflectionTest = TestCase (
-  testResondsTo = (
-    self assert: (Object new respondsTo: #isNil).
-    self assert: (23 respondsTo: #isNil).
-    self assert: (23 respondsTo: #+).
-  )
-  
-  testMethods = (
-    "First method in Object should be #class."
-    self assert: #class equals: (Object methods at: 1) signature.
-    self assert: (Object hasMethod: #==).
-  )
 
-  testPerform = (
-    | o |
-    self assert: Integer equals: (23 perform: #class).
-    self assert: (23 perform: #between:and: withArguments: (Array with: 22 with: 24)).
+    (*public testResondsTo = (
+      self assert: ((ObjectMirror reflecting: Object new) respondsTo: #isNil).
+      self assert: ((ObjectMirror reflecting: 23) respondsTo: #isNil).
+      self assert: ((ObjectMirror reflecting: 23) respondsTo: #+).
+    )*)
     
-    o := SuperTest new.
-    self assert: #super equals: (o perform: #something inSuperclass: SuperTestSuperClass).
-    
-    "Trying to see whether the stack in bytecode-based SOMs works properly"
-    self assert: #a equals: ((23 perform: #class) = Integer ifTrue: [#a] ifFalse: [#b]).
+    (*public testMethods = (
+      | o |
+      o := ReflectionTarget new.
+      (ObjectMirror reflecting: o) classMirror methods doIndexes: [ :i |
+        ('' + ((ObjectMirror reflecting: o) classMirror methods) at: i) println.
+      ]
 
-    self assert: 28 equals: 5 + (23 perform: #value).
-  )
-  
-  testInstVarAtAndPut = (
-    | tmp |
-    "Testing #at: and #at:put:"
-    tmp := Pair withKey: 3 andValue: 42.
-    
-    self assert: tmp key equals: (tmp instVarAt: 1).
-    
-    tmp instVarAt: 1 put: #foo.
-    self assert: #foo    equals: tmp key.
-  )
+      self assert: #class equals: (Object methods at: 1) signature.
+      self assert: (Object hasMethod: #==).
+    )*)
+
+    public testPerform = (
+      | o |
+
+      self assert: ((ObjectMirror reflecting: 23) perform: #between:and: withArguments: (Array with: 22 with: 24)).
+      
+      (*o := SuperTest new.
+      self assert: #super equals: (o perform: #something inSuperclass: SuperTestSuperClass).*)
+      
+      (*Trying to see whether the stack in bytecode-based SOMs works properly*)
+
+      o := ReflectionTarget new.
+      assert: ((ObjectMirror reflecting: o) perform: ('get' asSymbol)) = 0.
+      ((ObjectMirror reflecting: o) perform: #set: withArguments: (Array with: 5)).
+      assert: ((ObjectMirror reflecting: o) perform: #get) = 5.
+      self assert: 28 equals: 5 + ((ObjectMirror reflecting: 23) perform: #value).
+    )
+  ) : ( TEST_CONTEXT = () )
 )
